requirements.txt
requests
python-dotenv
ldap3